refactor: localize proxy upload routing
